Accuracy fix

+++ src/Engine/Core.cs
@@ -10,11 +10,11 @@ public class Core
 
     private readonly Dictionary<int, long> _depthCounts = new();
     
-    private readonly Dictionary<(long Ply, PlyOutcome Outcome), int> _outcomes = new();
+    private readonly Dictionary<(long Ply, PlyOutcome Outcome), long> _outcomes = new();
 
     public IReadOnlyDictionary<int, long> DepthCounts => _depthCounts;
 
-    public IReadOnlyDictionary<(long Ply, PlyOutcome Outcome), int> Outcomes => _outcomes;
+    public IReadOnlyDictionary<(long Ply, PlyOutcome Outcome), long> Outcomes => _outcomes;
     
     public void Initialise()
     {
